+++ utils/constants.ts
@@ -1,0 +1,12 @@
+export const SUBGRAPH_SYNC_SECONDS = 5000
+export const ETH_ZERO_ADDRESS = '0x0000000000000000000000000000000000000000'
+
+// The Graph GraphQL endpoint
+export const GRAPHQL_ENDPOINT = 'http://localhost:8000/subgraphs/name/adamazad/aqua'
+
+// Ganache EVM endpoint
+export const EVM_ENDPOINT = 'http://localhost:8545'
+
+// Time constants
+export const ONE_MINUTE = 60
+export const ONE_HOUR = ONE_MINUTE * 60
